refactor: Simplify Todos page component
The code changes in this commit refactor the Todos page component to simplify its structure. The previous implementation included unnecessary code for fetching data from an API and rendering a list of todos. This functionality has been removed as it is not currently needed. Instead, the component now renders a simple heading for the ToDo App.